% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf_fit.R
\name{build_brms_shell}
\alias{build_brms_shell}
\title{Construct a minimal brmsfit object for interoperability}
\usage{
build_brms_shell(stanfit, formula, data, stan_file, stan_data)
}
\description{
Construct a minimal brmsfit object for interoperability
}
\keyword{internal}
