% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_summary}
\alias{efficiency_summary}
\title{Get Efficiency Summary Table}
\usage{
efficiency_summary(object)
}
\arguments{
\item{object}{A gsf_fit object.}
}
\value{
A data frame with summary statistics for efficiency measures.
}
\description{
Creates a summary table of all efficiency measures similar to Table 2
in Kumbhakar and Tsionas (2021).
}
