% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{extract_delta_coefficients}
\alias{extract_delta_coefficients}
\title{Extract Delta Coefficients (Slack Determinants)}
\usage{
extract_delta_coefficients(object)
}
\arguments{
\item{object}{A gsf_fit object.}
}
\value{
A data frame with coefficient estimates.
}
\description{
Extract posterior estimates of the Delta coefficients that determine
input-specific slacks.
}
