% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{loo_cv}
\alias{loo_cv}
\title{LOO-CV for Model Comparison}
\usage{
loo_cv(object)
}
\arguments{
\item{object}{A gsf_fit object.}
}
\value{
A loo object.
}
\description{
Compute Leave-One-Out Cross-Validation using the loo package.
}
