% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf_fit.R
\name{parse_gsf_formula}
\alias{parse_gsf_formula}
\title{Parse GSF Formula}
\usage{
parse_gsf_formula(formula, data)
}
\arguments{
\item{formula}{The model formula}

\item{data}{The data frame}
}
\value{
A list with parsed components
}
\description{
Parse GSF Formula
}
\keyword{internal}
