% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_best_worst_firms}
\alias{plot_best_worst_firms}
\title{Compare Best and Worst Firms}
\usage{
plot_best_worst_firms(object, n_firms = 5, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{n_firms}{Number of best/worst firms to compare (default: 5).}

\item{...}{Additional arguments.}
}
\value{
A list of ggplot objects.
}
\description{
Creates comparison plots for the most and least efficient firms,
similar to Figure 3 in Kumbhakar and Tsionas (2021).
}
