% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_efficiency_distributions}
\alias{plot_efficiency_distributions}
\title{Plot Efficiency Distributions}
\usage{
plot_efficiency_distributions(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Efficiency Distributions
}
\keyword{internal}
