% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_efficiency_pairs}
\alias{plot_efficiency_pairs}
\title{Plot Efficiency Pairs}
\usage{
plot_efficiency_pairs(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Efficiency Pairs
}
\keyword{internal}
