% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_elasticities}
\alias{plot_elasticities}
\title{Plot Input Elasticities}
\usage{
plot_elasticities(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Input Elasticities
}
\keyword{internal}
