% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_input_slacks}
\alias{plot_input_slacks}
\title{Plot Input Slacks}
\usage{
plot_input_slacks(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Input Slacks
}
\keyword{internal}
