% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_marginal_effects}
\alias{plot_marginal_effects}
\title{Plot Marginal Effects of Z Variables}
\usage{
plot_marginal_effects(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plots the marginal effects of z variables on input slacks.
}
