% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_output_loss}
\alias{plot_output_loss}
\title{Plot Output Loss}
\usage{
plot_output_loss(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Output Loss
}
\keyword{internal}
