% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_returns_to_scale}
\alias{plot_returns_to_scale}
\title{Plot Returns to Scale}
\usage{
plot_returns_to_scale(object, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot Returns to Scale
}
\keyword{internal}
