% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot MCMC Trace Plots}
\usage{
plot_trace(object, pars = NULL, ...)
}
\arguments{
\item{object}{A gsf_fit object.}

\item{pars}{Character vector of parameter names to plot.}

\item{...}{Additional arguments.}
}
\value{
A plot.
}
\description{
Plot MCMC Trace Plots
}
\keyword{internal}
