% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{waic}
\alias{waic}
\title{WAIC for Model Comparison}
\usage{
waic(object)
}
\arguments{
\item{object}{A gsf_fit object.}
}
\value{
A list with WAIC value and effective number of parameters.
}
\description{
Compute the Widely Applicable Information Criterion (WAIC) for model
comparison.
}
